requirements.txt
fastapi
uvicorn
newspaper3k
duckduckgo_search
aiohttp
pydantic